Added app setup to a separate file
